+++ Changelog.md
@@ -6,6 +6,7 @@ AppleALC Changelog
 - Added Model key with possible Laptop/Desktop/Any values to Controllers.plist (matches IODT/comptaible)
 - Added VIA VT2020/2021 codec support layout-id 5, 7 Mirone resources
 - Added Z97 HDEF controller patch
+- Added Intel HD4000 HDMI support ig-platform-id 0x0166000A for Laptop and Desktop
 
 #### v1.0.5
 - Fixed AppleHDAController patching regression